+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat(control-tower): lz management api gov cloud support
 - chore(test): update all-enabled custom config rule lambda python version
 - fix(control-tower): exclude ignored ou from registering with control tower
+- fix(control-tower): manage ignored ou creation and registration
 
 ### Changed